+++ rotary/torch-ext/rotary/__init__.py
@@ -1,5 +1,5 @@
 from .ext import apply_rotary_transformers
-from .functions import apply_rotary
+from .layers import apply_rotary
 
 
 __all__ = ["apply_rotary", "apply_rotary_transformers"]
